requirements.txt
click==7.1.2
colorama==0.4.4
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
joblib==1.4.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
scikit-learn==1.4.2
scipy==1.13.0
shellingham==1.4.0
six==1.16.0
threadpoolctl==3.4.0
typer==0.3.2
